makefile: Unsilence silenced warnings

+++ util/vptr.c
@@ -15,7 +15,7 @@ bool vptr_equals(vptr p1, vptr p2)
 {return p1 == p2;}
 
 unsigned vptr_hash(vptr p)
-{return (unsigned)p;}
+{return (unsigned)(long long)p;}
 
 void vptr_print(vptr p, FILE* fp)
 {fprintf(fp, "%p", p);}
